Check if unindexed strings match the search and also add them to the list of matches.
