+++ pkg/web/documents.go
@@ -16,7 +16,6 @@ import (
 	"fmt"
 	"io"
 	"log/slog"
-	"maps"
 	"net/http"
 	"slices"
 	"strconv"
@@ -29,7 +28,6 @@ import (
 	"github.com/jackc/pgx/v5/pgxpool"
 
 	"github.com/ISDuBA/ISDuBA/pkg/database/query"
-	"github.com/ISDuBA/ISDuBA/pkg/itertools"
 	"github.com/ISDuBA/ISDuBA/pkg/models"
 )
 
@@ -543,250 +541,4 @@ func scanRows(
 		return nil, fmt.Errorf("scanning failed: %w", err)
 	}
 	return results, nil
-}
-
-// documentTexts is an end point to return JSON paths and highlighting positions.
-//
-//	@Summary		Returns List of JSON paths with highlighting positions.
-//	@Description	Returns List of JSON paths with highlighting positions inside text matches found by the search query.
-//	@Param			id		path	int	true	"Document ID"
-//	@Param			query	query	string	false	"Document query"
-//	@Param			include	query	bool	false	"Include the texts in the result"
-//	@Produce		json
-//	@Success		200	{object}	models.TextPaths
-//	@Failure		400	{object}	models.Error
-//	@Failure		401
-//	@Failure		500	{object}	models.Error
-//	@Router			/documents/texts/{id} [get]
-func (c *Controller) documentTexts(ctx *gin.Context) {
-	// Get an ID from context
-	docID, ok := parse(ctx, toInt64, ctx.Param("id"))
-	if !ok {
-		return
-	}
-
-	parser := query.Parser{
-		Mode:            query.DocumentMode,
-		MinSearchLength: MinSearchLength,
-	}
-
-	// The query to filter the documents.
-	expr, ok := parse(ctx, parser.Parse, ctx.DefaultQuery("query", "true"))
-	if !ok {
-		return
-	}
-
-	// Should the texts be included in the result?
-	include, ok := parse(ctx, strconv.ParseBool, ctx.DefaultQuery("include", "false"))
-	if !ok {
-		return
-	}
-
-	searches := slices.Collect(itertools.Apply(
-		expr.Searches(),
-		func(search string) string {
-			return query.LikeEscape(search)
-		},
-	))
-	if len(searches) == 0 {
-		// No search texts found in query -> no need to proceed.
-		ctx.JSON(http.StatusOK, models.TextPaths{})
-		return
-	}
-	ilikes, err := query.CompileILike(searches...)
-	if err != nil {
-		slog.Error("compiling ilikes failed", "err", err)
-		models.SendError(ctx, http.StatusInternalServerError, err)
-		return
-	}
-
-	searchTerm, replacements := query.
-		CreateTextSearchWhereClause("ut.txt", expr)
-	replacements = append(replacements, docID)
-
-	tlps := c.tlps(ctx)
-
-	const (
-		publisherTLPSQL = `` +
-			`SELECT` +
-			` ads.publisher,` +
-			` docs.tlp ` +
-			`FROM` +
-			` documents docs JOIN` +
-			` advisories ads ON docs.advisories_id = ads.id ` +
-			`WHERE docs.id = $1`
-		uniqueTextsSQL = `` +
-			`SELECT` +
-			` dt.num,` +
-			` ut.txt ` +
-			`FROM` +
-			` documents_texts dt JOIN` +
-			` unique_texts ut ON ut.id = dt.txt_id ` +
-			`WHERE` +
-			` dt.documents_id = $%d AND %s`
-		documentSQL = `` +
-			`SELECT document ` +
-			`FROM documents ` +
-			`WHERE id = $1`
-	)
-
-	//  Data to be loaded from the database.
-	var (
-		uniqueTexts  = make(map[int64]string)
-		documentData []byte
-	)
-
-	var forbidden bool
-	switch err := c.db.Run(
-		ctx.Request.Context(),
-		func(rctx context.Context, conn *pgxpool.Conn) error {
-			// Check permissions.
-			var publisher, tlp string
-			if err := conn.QueryRow(rctx, publisherTLPSQL, docID).Scan(
-				&publisher,
-				&tlp,
-			); err != nil {
-				return fmt.Errorf("failed to extract publisher and tlp: %w", err)
-			}
-			if len(tlps) > 0 && !tlps.Allowed(publisher, models.TLP(tlp)) {
-				forbidden = true
-				return nil
-			}
-			// Load unique texts.
-			query := fmt.Sprintf(uniqueTextsSQL, len(replacements), searchTerm)
-			rows, err := conn.Query(rctx, query, replacements...)
-			if err != nil {
-				return fmt.Errorf("loading unique texts failed: %w", err)
-			}
-			defer rows.Close()
-			for rows.Next() {
-				var (
-					textID int64
-					text   string
-				)
-				if err := rows.Scan(&textID, &text); err != nil {
-					return fmt.Errorf("scanning unique texts failed: %w", err)
-				}
-				uniqueTexts[textID] = text
-			}
-			if err := rows.Err(); err != nil {
-				return fmt.Errorf("scanning unique texts failed: %w", err)
-			}
-			if len(uniqueTexts) == 0 {
-				// No matches.
-				return nil
-			}
-			rows.Close()
-			// Load the document itself.
-			if err := conn.QueryRow(rctx, documentSQL, docID).Scan(&documentData); err != nil {
-				return fmt.Errorf("loading document failed: %w", err)
-			}
-			return nil
-		},
-		c.cfg.Database.MaxQueryDuration, // In case the user provided a very expensive query.
-	); {
-	case errors.Is(err, pgx.ErrNoRows):
-		models.SendErrorMessage(ctx, http.StatusNotFound, "document not found")
-		return
-	case err != nil:
-		slog.Error("database error", "err", err)
-		models.SendError(ctx, http.StatusInternalServerError, err)
-		return
-	case forbidden:
-		models.SendErrorMessage(ctx, http.StatusForbidden, "access denied")
-		return
-	}
-
-	if len(uniqueTexts) == 0 {
-		ctx.JSON(http.StatusOK, models.TextPaths{})
-		return
-	}
-
-	var document any
-	dec := json.NewDecoder(bytes.NewReader(documentData))
-	// We use Numbers here to make it easier to
-	// filter out strings that apparently are not IDs like 3.1415
-	// which would be otherwise parse as float64s. 3.000 would
-	// also be parsed into float64s as would 3. Not using Numbers
-	// would make hard to tell them apart.
-	dec.UseNumber()
-	if err := dec.Decode(&document); err != nil {
-		models.SendError(ctx, http.StatusInternalServerError, err)
-		return
-	}
-	paths := buildTextPaths(document, uniqueTexts, ilikes, include)
-	ctx.JSON(http.StatusOK, paths)
-}
-
-// There are integers in CSAF documents which are not text ids.
-var knownNoneTexts = map[string]struct{}{
-	"/document/tracking/version": {},
-	// TODO: Fill me!
-}
-
-// buildTextPaths traverses the document tree and attemps to resolve
-// the unique texts producing a list of matches.
-func buildTextPaths(
-	document any,
-	uniqueTexts map[int64]string,
-	ilikes query.ILikeExpr,
-	include bool,
-) models.TextPaths {
-	paths := make(models.TextPaths, 0, len(uniqueTexts))
-
-	store := func(path []string, id int64) {
-		joined := strings.Join(path, "")
-		// Ignore the ids known to be regular texts.
-		if _, ok := knownNoneTexts[joined]; ok {
-			return
-		}
-		// Only include this text if we have text
-		// with this id.
-		if text, ok := uniqueTexts[id]; ok {
-			var t *string
-			if include {
-				t = &text
-			}
-			paths = append(paths, models.TextPath{
-				Path:      joined,
-				Text:      t,
-				Positions: ilikes.Search(text),
-			})
-		}
-	}
-
-	// Recursively traverse the document and find the text ids.
-	var recurse func(any, []string)
-	recurse = func(curr any, path []string) {
-		switch x := curr.(type) {
-		case map[string]any:
-			// Sort to make output deterministic.
-			keys := slices.Sorted(maps.Keys(x))
-			for _, key := range keys {
-				path = append(path, "/", key)
-				recurse(x[key], path)
-				path = path[:len(path)-2]
-			}
-		case []any:
-			for i, y := range x {
-				path = append(path, "["+strconv.Itoa(i)+"]")
-				recurse(y, path)
-				path = path[:len(path)-1]
-			}
-		case json.Number:
-			id, err := x.Int64()
-			if err != nil {
-				return
-			}
-			store(path, id)
-		case string:
-			id, err := strconv.ParseInt(x, 10, 64)
-			if err != nil {
-				return
-			}
-			store(path, id)
-		}
-	}
-	recurse(document, nil)
-	return paths
 }
